+++ docs/versionhistory.rst
@@ -5,6 +5,9 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
 **UNRELEASED**
 
+- Added context manager mix-in classes (``anyio.ContextManagerMixin`` and
+  ``anyio.AsyncContextManagerMixin``) to help write classes that embed other context
+  managers (particularly cancel scopes or task groups)
 - Added the ability to specify the thread name in ``start_blocking_portal()``
   (`#818 <https://github.com/agronholm/anyio/issues/818>`_; PR by @davidbrochart)